Slackbot improvements for scanning AWS accounts
Now it is possible to specify additional parameters to bot, such as: regions,
security features and tags.
Note, I haven't found a proper way to make scan conversation with bot work in *both*
direct channel and group chanel in slack using slackbot package. This should be reimplemented
and improved in the future.