Add workaround for hammer bot "status" command
This message appears in logs when issuing "status" command:
"lost websocket connection, try to reconnect now".
The problem appears when the "status" message is sent to slack. It appears to be
long message (althoug it's less then 40000 chars slack limit) and the slack should
just truncate the message instead of closing the connection. But ii works if we use
shorter messages.